GuestInfo 파일 수정
