+++ sublet-front/src/components/FetchList.js
@@ -102,21 +102,14 @@ async function FetchReservation(setReservationInfo) {
   }, []);
 }
 
-async function FetchReservationByPostKey(post_key) {
-  const [reservationInfo, setReservationInfo] = useState([]);
+async function FetchReservationByPostKey(setReservationInfo, postKey) {
   const URL =
-    `${process.env.REACT_APP_BACKEND_URL}/reservation/post?key=` + post_key;
+    `${process.env.REACT_APP_BACKEND_URL}/reservation/post?key=` + postKey;
 
-  const getPostInfo = async () => {
-    const json = await fetch(URL, headerOptions('GET'))
-      .then(notFoundError)
-      .catch(raiseError('FetchReservationByPostKey'));
-    setReservationInfo(json);
-  };
-
-  useEffect(() => {
-    getPostInfo();
-  }, []);
+  const json = await fetch(URL, headerOptions('GET'))
+    .then(notFoundError)
+    .catch(raiseError('FetchReservationByPostKey'));
+  setReservationInfo(json);
 
   const reservation = Array.from(reservationInfo);
   return reservation;
@@ -184,22 +177,13 @@ async function FetchImage(formData) {
   });
 }
 
-async function FetchGetMyUser() {
-  const [userInfo, setUserInfo] = useState();
+async function FetchGetMyUser(setUserInfo) {
   const URL = `${process.env.REACT_APP_BACKEND_URL}/user/profile`;
 
-  const getUserInfo = async () => {
-    const json = await fetch(URL, headerOptions('GET'))
-      .then(notFoundError)
-      .catch(raiseError('FetchGetMyUser'));
-    setUserInfo(json);
-  };
-
-  useEffect(() => {
-    getUserInfo();
-  }, []);
-
-  return userInfo;
+  const json = await fetch(URL, headerOptions('GET'))
+    .then(notFoundError)
+    .catch(raiseError('FetchGetMyUser'));
+  setUserInfo(json);
 }
 
 async function FetchGetOneUser(userId) {
