+++ sublet-front/src/pages/GuestInfo/components/Info/VerifyList.js
@@ -1,0 +1,9 @@
+import { NormalText } from '@components/styles/Public.styles';
+
+export const VerifyList = (k, v) => {
+  return (
+    <NormalText>
+      {k} {v ? '인증 완료✅' : '인증 안됨❌'}
+    </NormalText>
+  );
+};
