+++ sublet-front/src/pages/GuestInfo/components/UserProfile/PostListComponent.js
@@ -1,11 +1,11 @@
-import { PostSummaryBlock } from '../SummaryBlock';
-import { DateFormat, priceToString } from '../StaticComponents.js';
-import * as s from '../styles/Public.styles.js';
-import { FetchGetPost } from '../FetchList';
 import { guestInfoPopUpStore } from '../store/guestInfoStore.js';
 import { useState } from 'react';
+import { NormalButton, SecondHead } from '@components/styles/Public.styles';
+import { PostSummaryBlock } from './Blocks/PostSummaryBlock';
+import { DateFormat, priceToString } from '@components/StaticComponents';
+import { FetchGetPost } from '@components/FetchList.js';
 
-function PostInfo(user_id) {
+function PostListComponent(user_id) {
   const [postInfo, setPostInfo] = useState([]);
   FetchGetPost(user_id.user_id, setPostInfo);
 
@@ -15,7 +15,7 @@ function PostInfo(user_id) {
 
   return (
     <div className="mb-4 mt-8">
-      <s.SecondHead className="inline">방 현황</s.SecondHead>
+      <SecondHead className="inline">방 현황</SecondHead>
       {postInfo.length > 0 ? (
         postInfo.map(res => {
           const address = res.city + ' ' + res.gu + ' ' + res.dong;
@@ -35,9 +35,9 @@ function PostInfo(user_id) {
       ) : (
         <p className="text-base font-extrabold">올린 방이 아직 없습니다.</p>
       )}
-      <s.NormalButton onClick={setPostPopUpState}>방 올리기</s.NormalButton>
+      <NormalButton onClick={setPostPopUpState}>방 올리기</NormalButton>
     </div>
   );
 }
 
-export { PostInfo };
+export { PostListComponent };
